iviewer-tobot
Install iviewer-robot
Hello and welcome to the guide to working with iviewer-Robot In this article, we teach you how to install this tool

About  iviewer-robot
This tool is an open source tool and anyone can participate in its development This tool is actually a robot that works with a loop algorithm and can do a lot of work. In fact, if this robot does what a human can do, but in higher numbers. This is the initial version of this tool and is still under development. Software engineers can work together to develop these tools and be honest in their development.

Language Programming
Python 3.8.5
Python Library Selenium

About Me
I am a 13-year-old software engineer and I do not like to be limited to one branch of IT and I like to work in fields such as network engineering and security and artificial intelligence engineering. In short, I am a 13-year-old programmer who enjoys books and training. I use to learn
Installation and use iviewer-robot
Step 1: > Install Language Python 3
To install Python, open the terminal or CMD and type this phrase
For Linux Terminal
apt-get install python3
For Windows
Download Python from Python site

Step 2: > Download Python-pip3
For Linux Terminal
Sudo apt-get install python3-pip
For Winsows
By downloading Python, the pipe will be installed

Step 3: > Install TB Library
For Linux
Pip3 install selenium or pip install selenium
For Windown
Pip3 install selenium or pip install selenium

Step 4: >Install Chrome Driver
Download the Chrome driver compatible with the Chrome browser installed on the system

Step 4: >Download iviewer-robot
Link Download In Github
https://codeload.github.com/Aryia-Behroziuan/iviewer-robot/zip/main

Version iviewer-robot
iviewer-robot 1.0 = Aparat-robot
iviewer-robot 2.0 =  YouTube-robot

Step 4: >Run iviewer-robot in Terminal or CMD 
Python3 iviewer-robot 1.0.py
Or 
Python iviewer-robot 1.0 or 2.0.py



GOODBOY